Reorganize screen views and their wrappers
