+++ app/screens/MapViewContainer.js
@@ -3,7 +3,7 @@ import getLocationToSave from '../actions/action_dropNewPin.js';
 import getPins from '../actions/action_updatePins.js';
 import deletePin from '../actions/action_deletePin.js';
 import updateRecent from '../actions/action_updateRecent.js';
-import PinListView from '../components/PinListView.js';
+import MapView from './MapView.js';
 
 function mapStateToProps(state) {
   return {
@@ -12,4 +12,4 @@ function mapStateToProps(state) {
   };
 }
 
-export default connect(mapStateToProps, { getLocationToSave, getPins, deletePin, updateRecent })(PinListView);
+export default connect(mapStateToProps, { getLocationToSave, getPins, deletePin, updateRecent })(MapView);
